% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{material_card}
\alias{material_card}
\title{helpers}
\usage{
material_card(..., header = NULL, bgcolor = "#191a1a")
}
\description{
utils functions
material_card

provides a card for major UI elements
}
